% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{CityWeather}
\alias{CityWeather}
\title{CityWeather object}
\usage{
CityWeather(
  city,
  temperature = NA_real_,
  humidity = NA_real_,
  pressure = NA_real_
)
}
\arguments{
\item{city}{Character, name of the city.}

\item{temperature}{Numeric, temperature in degrees Celsius.}

\item{humidity}{Numeric, relative humidity in percent (0–100).}

\item{pressure}{Numeric, atmospheric pressure in hPa.}
}
\value{
An object of class \code{"City_Weather"}.
}
\description{
Small S3 class that stores basic weather information for one city:
temperature (°C), humidity (\%) and pressure (hPa).
}
\examples{
cw <- CityWeather("Paris", temperature = 12, humidity = 85, pressure = 1015)
cw

}
