% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{GlassProfile}
\alias{GlassProfile}
\title{Create a Glass Profile object}
\usage{
GlassProfile(a, b, x_1, x_2, x_3, x_4, r_foot, r_stem, r_bowl, r_rim)
}
\arguments{
\item{a, b}{Numeric, lower and upper integration bounds (cm).}

\item{x_1, x_2, x_3, x_4}{Breakpoints of the radius profile.}

\item{r_foot, r_stem, r_bowl, r_rim}{Radii of the different parts of the glass.}
}
\value{
An object of class Glass_Profile.
}
\description{
Minimal S3 class storing geometry of the champagne glass.
}
