% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{print.City_Weather}
\alias{print.City_Weather}
\title{Print method for City_Weather}
\usage{
\method{print}{City_Weather}(x, ...)
}
\arguments{
\item{x}{A \code{City_Weather} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Nicely prints basic weather information for a city.
}
