% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r}
\alias{r}
\title{Radius profile of the champagne glass}
\usage{
r(t, x_1, x_2, x_3, x_4, r_foot, r_stem, r_bowl, r_rim)
}
\arguments{
\item{t}{Position along the vertical axis (cm).}

\item{x_1, x_2, x_3, x_4}{Breakpoints (cm) that define foot, stem, bowl and rim transitions.}

\item{r_foot, r_stem, r_bowl, r_rim}{Radii (cm) at foot, stem, bowl and rim.}
}
\value{
Radius at height t (cm).
}
\description{
Piecewise radius function r(t) defining the shape of the glass.
}
\examples{
r(10, x_1 = 1, x_2 = 5, x_3 = 10, x_4 = 15,
  r_foot = 5, r_stem = 1, r_bowl = 6, r_rim = 5)
}
