% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r_vec_for_integrate}
\alias{r_vec_for_integrate}
\title{Vectorised wrapper for integrate()}
\usage{
r_vec_for_integrate(t, x)
}
\arguments{
\item{t}{Numeric vector of positions.}

\item{x}{A Glass_Profile object.}
}
\value{
Numeric vector of radii.
}
\description{
Vectorised wrapper for integrate()
}
