% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{radius_cone.Glass_Profile}
\alias{radius_cone.Glass_Profile}
\title{Radius at height t}
\usage{
\method{radius_cone}{Glass_Profile}(x, t, ...)
}
\arguments{
\item{x}{A Glass_Profile object.}

\item{t}{Numeric height value.}
}
\value{
The radius (cm) at height t.
}
\description{
S3 method for Glass_Profile class.
}
