% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{radius_cone}
\alias{radius_cone}
\alias{radius_cone.Glass_Profile}
\title{Radius at a given level}
\usage{
radius_cone(x, ...)

\method{radius_cone}{Glass_Profile}(x, t, ...)
}
\arguments{
\item{x}{A Glass_Profile object.}

\item{...}{Further arguments (currently unused).}

\item{t}{Numeric level along the glass axis.}
}
\value{
Numeric radius at level \code{t}.
}
\description{
Generic and method for champagne glass profiles.
}
\examples{
glass <- GlassProfile(
  a = 12.6, b = 22.5,
  x_1 = 1.1, x_2 = 12.6, x_3 = 18.8, x_4 = 22.5,
  r_foot = 5.8, r_stem = 0.25, r_bowl = 7.5, r_rim = 7
)
radius_cone(glass, t = 14)
}
