% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{S}
\alias{S}
\title{Smooth easing function S}
\usage{
S(t)
}
\arguments{
\item{t}{Numeric scalar, typically between 0 and 1.}
}
\value{
Numeric scalar between 0 and 1.
}
\description{
This function implements a cosine-based easing that smoothly goes
from 0 to 1 as t increases from 0 to 1.
}
