% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{volume_between.Glass_Profile}
\alias{volume_between.Glass_Profile}
\title{Volume of champagne between two heights}
\usage{
\method{volume_between}{Glass_Profile}(x, a = x$a, b = x$b, ...)
}
\arguments{
\item{x}{A Glass_Profile object.}

\item{a, b}{Heights (cm).}
}
\value{
Volume in cm³.
}
\description{
Computes volume using the disk method.
}
