% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\name{volume_between}
\alias{volume_between}
\alias{volume_between.Glass_Profile}
\title{Volume between two levels}
\usage{
volume_between(x, ...)

\method{volume_between}{Glass_Profile}(x, a = x$a, b = x$b, ...)
}
\arguments{
\item{x}{A Glass_Profile object.}

\item{...}{Further arguments (currently unused).}

\item{a, b}{Lower and upper fill levels.}
}
\value{
Numeric volume in cubic centimetres.
}
\description{
Compute champagne volume between fill levels \code{a} and \code{b}.
}
\examples{
glass <- GlassProfile(
  a = 12.6, b = 22.5,
  x_1 = 1.1, x_2 = 12.6, x_3 = 18.8, x_4 = 22.5,
  r_foot = 5.8, r_stem = 0.25, r_bowl = 7.5, r_rim = 7
)
volume_between(glass)
}
