% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r}
\alias{r}
\title{Piecewise radius function for the champagne glass}
\usage{
r(t, x_1, x_2, x_3, x_4, r_foot, r_stem, r_bowl, r_rim)
}
\arguments{
\item{t}{Position along the horizontal axis.}

\item{x_1, x_2, x_3, x_4}{Breakpoints.}

\item{r_foot, r_stem, r_bowl, r_rim}{Radii of the different parts (cm).}
}
\value{
Radius at position t (cm).
}
\description{
Piecewise radius function for the champagne glass
}
