% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r_vec_for_integrate}
\alias{r_vec_for_integrate}
\title{Vectorized wrapper of r() for integration}
\usage{
r_vec_for_integrate(t, ...)
}
\arguments{
\item{t}{Numeric vector of heights.}

\item{...}{Passed to r().}
}
\value{
Numeric vector of radii.
}
\description{
Adapts r() to work with integrate(), which passes vector t.
}
\examples{
r_vec_for_integrate(c(10, 11),
  x_1 = 1, x_2 = 5, x_3 = 10, x_4 = 15,
  r_foot = 5, r_stem = 1, r_bowl = 6, r_rim = 5)
}
