% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{S}
\alias{S}
\title{Easing function S(t) on \link{0, 1}}
\usage{
S(t)
}
\arguments{
\item{t}{Numeric in \link{0, 1}.}
}
\value{
Numeric, smoothed value between 0 and 1.
}
\description{
Easing function S(t) on \link{0, 1}
}
